fix: `docs.rs` builds breaking
Currently, `docs.rs` builds for this repo [fail][docsrs-failure];
pulling out the relevant error snippet:

```
[INFO] [stderr] error: failed to run custom build command for `opentelemetry-stackdriver v0.6.0 (/opt/rustwide/workdir)`
[INFO] [stderr]
[INFO] [stderr] Caused by:
[INFO] [stderr]   process didn't exit successfully: `/opt/rustwide/target/debug/build/opentelemetry-stackdriver-5bb9eeda67370cd7/build-script-build` (exit code: 101)
[INFO] [stderr] --- stdout
[INFO] [stderr] out: Output { status: ExitStatus(ExitStatus(256)), stdout: "", stderr: "error: \'rustfmt\' is not installed for the toolchain \'nightly-x86_64-unknown-linux-gnu\'\nTo install, run `rustup component add rustfmt --toolchain nightly-x86_64-unknown-linux-gnu`\n" }
```

[docsrs-failure]: https://docs.rs/crate/opentelemetry-stackdriver/0.6.0/builds/259476

This is because the current `build.rs` in `master` attempts to run
`rustfmt` as part of the default behavior for `tonic-build` with default
features enabled:

https://github.com/vivint-smarthome/opentelemetry-stackdriver/blob/52d242073e04091b3b3621f18e66fde9f6e719af/build.rs#L2-L13

...and `docs.rs` does not include `rustfmt` in its Rust toolchain
installation (justifiably so, since it probably _shouldn't_ have any
interesting in mutating code it builds docs for).

[`docs.rs`' Build page](https://docs.rs/about/builds) suggests that
`build.rs` scripts use the `DOCS_RS` runtime variable:

> To recognize Docs.rs from `build.rs` files, you can test for the
> environment variable `DOCS_RS`, e.g.:
>
>     if let Ok(_) = std::env::var("DOCS_RS") {
>         // ... your code here ...
>     }

...which is what is implemented here. This _could_ also be handled by
separating the generation of code from `build.rs`, pushing the
responsibility to developers, and having a CI check for up-to-date
codegen. There can be [good trade-offs][manual-codegen-tradeoffs] with
this, but methinks that that decision can and should be separate from
the easy resolution above.

[manual-codegen-tradeoffs]: https://epage.github.io/blog/2019/10/speeding-up-rust-builds-code-gen-edition/