squash! docs: add a step-by-step example for top-level lib documentation
New commit message:

This shouldn't be the place where documentation stops being written
(which is being discussed in #7), but it might be a good start. This is
very similar to code you can already find in this repo's
[`examples/basic.rs`][basic-example] and documentation in other places,
like [the doc example for
`tracing_opentelemetry::OpenTelemetryLayer::new`][open_telemetry_layer-new].
The focus of these docs, though, is to demonstrate the usage of the API
in an incremental style, so that single steps are as easy as possible to
understand.

[basic-example]: https://github.com/vivint-smarthome/opentelemetry-stackdriver/blob/102ab6242a891798a8316db5f7bfca463c818507/examples/basic.rs
[open_telemetry_layer-new]: https://docs.rs/tracing-opentelemetry/0.8.0/tracing_opentelemetry/struct.OpenTelemetryLayer.html#examples